requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --extra=dev
#
appdirs==1.4.4
    # via black
attrs==21.2.0
    # via pytest
black==21.5b1
    # via biopsias (setup.py)
blis==0.7.4
    # via
    #   spacy
    #   thinc
catalogue==2.0.4
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   black
    #   nltk
    #   pip-tools
    #   sacremoses
    #   typer
cycler==0.10.0
    # via matplotlib
cymem==2.0.5
    # via
    #   preshed
    #   spacy
    #   thinc
datasets==1.8.0
    # via biopsias (setup.py)
dill==0.3.4
    # via
    #   datasets
    #   multiprocess
filelock==3.0.12
    # via
    #   huggingface-hub
    #   transformers
flake8==3.9.2
    # via biopsias (setup.py)
fsspec==2021.6.1
    # via datasets
fuzzywuzzy==0.18.0
    # via biopsias (setup.py)
huggingface-hub==0.0.8
    # via
    #   datasets
    #   transformers
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
isort==5.8.0
    # via biopsias (setup.py)
jinja2==3.0.1
    # via spacy
joblib==1.0.1
    # via
    #   nltk
    #   sacremoses
    #   scikit-learn
kiwisolver==1.3.1
    # via matplotlib
markupsafe==2.0.1
    # via jinja2
matplotlib==3.4.2
    # via biopsias (setup.py)
mccabe==0.6.1
    # via flake8
multiprocess==0.70.12.2
    # via datasets
murmurhash==1.0.5
    # via
    #   preshed
    #   spacy
    #   thinc
mypy-extensions==0.4.3
    # via black
nltk==3.6.2
    # via biopsias (setup.py)
numpy==1.20.3
    # via
    #   biopsias (setup.py)
    #   blis
    #   datasets
    #   matplotlib
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   seqeval
    #   spacy
    #   thinc
    #   torch
    #   transformers
packaging==20.9
    # via
    #   datasets
    #   pytest
    #   spacy
    #   transformers
pandas==1.2.4
    # via
    #   biopsias (setup.py)
    #   datasets
pathspec==0.8.1
    # via black
pathy==0.5.2
    # via spacy
pep517==0.10.0
    # via pip-tools
pillow==8.2.0
    # via matplotlib
pip-tools==6.1.0
    # via biopsias (setup.py)
pluggy==0.13.1
    # via pytest
preshed==3.0.5
    # via
    #   spacy
    #   thinc
py==1.10.0
    # via pytest
pyarrow==3.0.0
    # via datasets
pycodestyle==2.7.0
    # via flake8
pydantic==1.7.4
    # via
    #   spacy
    #   thinc
pyflakes==2.3.1
    # via flake8
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pytest==6.2.4
    # via biopsias (setup.py)
python-dateutil==2.8.1
    # via
    #   matplotlib
    #   pandas
python-levenshtein==0.12.2
    # via biopsias (setup.py)
pytokenizations==0.8.3
    # via biopsias (setup.py)
pytz==2021.1
    # via pandas
rapidfuzz==1.4.1
    # via spaczz
regex==2021.4.4
    # via
    #   black
    #   nltk
    #   sacremoses
    #   spaczz
    #   transformers
requests==2.25.1
    # via
    #   datasets
    #   huggingface-hub
    #   spacy
    #   transformers
sacremoses==0.0.45
    # via transformers
scikit-learn==0.24.2
    # via
    #   biopsias (setup.py)
    #   seqeval
scipy==1.6.3
    # via scikit-learn
seqeval==1.2.2
    # via biopsias (setup.py)
six==1.16.0
    # via
    #   cycler
    #   python-dateutil
    #   sacremoses
smart-open==3.0.0
    # via pathy
spacy-legacy==3.0.5
    # via spacy
spacy==3.0.6
    # via
    #   biopsias (setup.py)
    #   spaczz
spaczz==0.5.3
    # via biopsias (setup.py)
srsly==2.4.1
    # via
    #   spacy
    #   thinc
thinc==8.0.3
    # via spacy
threadpoolctl==2.1.0
    # via scikit-learn
tokenizers==0.10.3
    # via transformers
toml==0.10.2
    # via
    #   black
    #   pep517
    #   pytest
torch==1.8.1
    # via biopsias (setup.py)
tqdm==4.49.0
    # via
    #   datasets
    #   huggingface-hub
    #   nltk
    #   sacremoses
    #   spacy
    #   transformers
transformers==4.6.1
    # via biopsias (setup.py)
typer==0.3.2
    # via
    #   pathy
    #   spacy
typing-extensions==3.10.0.0
    # via torch
urllib3==1.26.5
    # via requests
wasabi==0.8.2
    # via
    #   spacy
    #   thinc
xlrd==2.0.1
    # via biopsias (setup.py)
xxhash==2.0.2
    # via datasets

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
